Creando condicional para eliminar

+++ app/Http/Controllers/PostController.php
@@ -178,6 +178,14 @@ class PostController extends Controller
      */
     public function delete(Post $post){
 
+
+      //Comparamos usuario mediante el Facade si es diferente
+      //no puede eliminar el post y lo redirige hacia al index
+      if($post->user_id != \Auth::user()->id){
+
+        return redirect()->route('posts_path');
+      }
+
        $post->delete();
 
       return redirect()->route('posts_path');
